Fix Bank And tag issue

+++ expense-api/src/main/java/com/expense/expensemanagement/service/tag/TagService.java
@@ -2,11 +2,15 @@ package com.expense.expensemanagement.service.tag;
 
 import com.expense.expensemanagement.conversion.EntityModalConversion;
 import com.expense.expensemanagement.dao.TagDAO;
+import com.expense.expensemanagement.dao.TagMappingDAO;
 import com.expense.expensemanagement.entity.Tag;
 import com.expense.expensemanagement.model.ResponseList;
 import com.expense.expensemanagement.model.TagModel;
+import com.expense.expensemanagement.service.tagMapping.TagMapping;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -21,11 +25,12 @@ import java.util.stream.Collectors;
 public class TagService implements ITagService {
 
     private final TagDAO tagDAO;
-
-
     @Autowired
     @Qualifier("TagEntityModel")
     private EntityModalConversion<Tag, TagModel> tagEntityModel;
+
+    @Autowired
+    private TagMappingDAO tagMappingDAO;
 
     @Autowired
     public TagService(TagDAO tagDAO ) {
@@ -46,13 +51,21 @@ public class TagService implements ITagService {
         Tag tag = tagDAO.findById(id).orElseThrow(
                 () -> new NoSuchElementException("Tag Not found")
         );
+        if(this.tagMappingDAO.findByTags(tag).size() != 0){
+            throw new IllegalStateException("Tag is being in used");
+        }
         this.tagDAO.deleteById(tag.getId());
     }
 
     public TagModel addTag(TagModel tagModel) {
-        Tag tagEntity = this.tagDAO.save(tagEntityModel.getEntity(tagModel));
-        tagModel.setId(tagEntity.getId());
-        return tagModel;
+        try {
+            Tag tagEntity = this.tagDAO.save(tagEntityModel.getEntity(tagModel));
+            tagModel.setId(tagEntity.getId());
+            return tagModel;
+        }
+        catch (DataIntegrityViolationException ex){
+            throw new DuplicateKeyException("name:Tag already exits. Tag name should be unique");
+        }
     }
 
     public List<Tag> findAllByIds(List<Long> ids) {
@@ -62,5 +75,10 @@ public class TagService implements ITagService {
             tagEntities.add(tagIterator.next());
         }
         return tagEntities;
+    }
+
+    @Override
+    public TagModel findTags(long tagId, String userId) {
+        return this.tagEntityModel.getModel(this.tagDAO.findByUserIdAndId(userId, tagId).orElseThrow(() -> new NoSuchElementException("Tag id not found for the user")));
     }
 }
